Add Uploaded module
