+++ app/jobs/upload_job.rb
@@ -1,10 +1,19 @@
-class UploadJob < ApplicationJob
-  queue_as :dradis_upload
+class UploadJob
+  include Resque::Plugins::Status
 
-  def perform(default_user_id:, file:, plugin_name:, project_id:, state:, uid:)
+  @queue = :dradis_upload
+
+  def perform
+    default_user_id = options['default_user_id']
+    file = options['file']
+    plugin_name = options['plugin_name']
+    project_id = options['project_id']
+    state = options['state']
+    uid = options['uid']
+
     logger = Log.new(uid: uid)
 
-    logger.write { "Job id is #{job_id}." }
+    logger.write { "Job id is #{uid}." }
     logger.write { 'Running Ruby version %s' % RUBY_VERSION }
     logger.write { 'Worker process starting background task.' }
 
@@ -21,5 +30,17 @@ class UploadJob < ApplicationJob
     importer.import(file: file)
 
     logger.write { 'Worker process completed.' }
+
+    completed
+  rescue => exception
+    logger.write { "There was an error with the upload: #{exception.message}" }
+    if Rails.env.development?
+      exception.backtrace.first(10).each do |trace|
+        logger.debug { trace }
+        sleep(0.2)
+      end
+    end
+
+    failed('message' => "There was an error with the upload: #{exception.message}")
   end
 end
