+++ engines/dradis-api/app/views/dradis/ce/api/v3/upload/create.json.jbuilder
@@ -1,1 +1,5 @@
-json.job_id @job_id
+if @job_id
+  json.job_id @job_id
+end
+
+json.status @status
