+++ engines/dradis-api/app/views/dradis/ce/api/v3/upload/show.json.jbuilder
@@ -1,0 +1,5 @@
+json.status @status
+
+if @message
+  json.message @message
+end
